requirements.txt
click==7.1.2
dominate==2.5.1
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-CKEditor==0.4.4.1
Flask-MySQLdb==0.2.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysqlclient==2.0.1
PyYAML==5.3.1
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.3